chore: renamed ExperimentContent to ExperimentAdapter
