+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/Controllers/ExperimentContent.html
@@ -105,7 +105,7 @@ var activeTableTab = "activeTableTab";
 <li>android.widget.ArrayAdapter&lt;<a href="../../../../com/cmput301w21t06/crowdfly/Models/Experiment.html" title="class in com.cmput301w21t06.crowdfly.Models">Experiment</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.Controllers.ExperimentContent</li>
+<li>com.cmput301w21t06.crowdfly.Controllers.ExperimentAdapter</li>
 </ul>
 </li>
 </ul>
